Stronghold - Rename: rateLimitState -> rateLimitStateModel
